chore: rename workflow file
